ColumnInfo Annotation support and multiple fixes (#1)
ColumnInfo Support

@ColumnInfo annotation in (Base-)Entities are converted to the drift named() syntax.
renamed annotation_generator.dart to class_generator.dart

was confusing with annotation_converter existing too.
used typeChecker in annotation_converter.dart instead of class name

Used actual type checker class to check which annotation to convert instead of just checking the string name of the class.
multiple Entities in one dao should work correctly

When multiple entities are used in one dao the tables parameter of the @DriftAccessor annotation is not being filled correctly with all used entities.
fixed multiple problems when the casing of the table name in a query is not correct

In various places in the generated code the casing of the table name was incorrect, if the table name was not exactly like the ClassName.
Additionally  @ColumnInfo didn't work if the casing wasn't exactly right.